pyproject.toml
[tool.poetry]
name = "tiny-telematics"
version = "0.1.0"
description = ""
authors = ["Christian Hollinger <christian@chollinger.com>"]

[tool.poetry.dependencies]
python = "3.8.13"
gps = "^3.19"
redis = "^4.3.4"
kafka-python = "^2.0.2"
jsonpickle = "^2.2.0"
geopy = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

